Added EB APs and it should work in out hallways now.  Modified to hopefully speed up performance. Also added gesturing to make Gordie happy!

+++ trunk/src/com/iDocent/WeightedScanFactory.java
@@ -13,7 +13,7 @@ public class WeightedScanFactory {
 	
 	public WeightedScan Create(String mac)
 	{
-		List<Integer> loc = locations.GetLocation(mac);
+		List<Float> loc = locations.GetLocation(mac);
 		
 		if(loc != null)
 			return new WeightedScan(mac, loc);
@@ -23,10 +23,10 @@ public class WeightedScanFactory {
 
 	//Get a demo weighted scan that bypasses the map for location
 	public WeightedScan Create(String mac, boolean demo) {
-		List<Integer> loc = new LinkedList<Integer>();
-		loc.add(0);
-		loc.add(0);		
-		loc.add(0);
+		List<Float> loc = new LinkedList<Float>();
+		loc.add(0f);
+		loc.add(0f);		
+		loc.add(0f);
 		return new WeightedScan(mac, loc);
 	}
 }
